feat(kcd): consolidate Sanity import scripts
